Added endpoint for adding games to users library

+++ Spring/src/main/java/pl/ttsw/GameRev/repository/WebsiteUserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
 @Repository
-public interface WebsiteUserRepository extends JpaRepository<WebsiteUser, Integer> {
+public interface WebsiteUserRepository extends JpaRepository<WebsiteUser, Long> {
     WebsiteUser findByUsername(String username);
     WebsiteUser findByUsernameOrEmail(String username, String email);
     WebsiteUser findByNickname(String nickname);
